Retrieve Articles with Favorite Status
Summary
The resolver checks if the articleId exists in the favourite_articles array of the user's document in the database.
If the articleId is found in the user's favourite_articles array, the isFavorite field in the response is set to true.
If the articleId is not found, isFavorite is set to false.

How to test
Toggle favaourite article
mutation {
  toggleFavouriteArticle(userId: "6713c70982539a6394424088", articleId: "671345982b9e701fbf58706c") {
    message
  }
}

Then,

query {
  getArticle(id: "671345982b9e701fbf58706c", userId: "6713c70982539a6394424088") {
    article_name
    article_desc
    article_genre
    isFavorite
  }
}

When we test it again, it shows the different result